lots of changes
refactored 64 bit integers to int32_t, libgd can handle that better,
they are more than enough and we prevent lots of typecasting, also
first code for ADF rendering

+++ ansilove/ansilove.h
@@ -66,6 +66,6 @@ typedef struct {
 sauce *sauceReadFileName(char *fileName);
 sauce *sauceReadFile(FILE *file);
 void  readRecord(FILE *file, sauce *record);
-void  readComments(FILE *file, char **comment_lines, int64_t comments);
+void  readComments(FILE *file, char **comment_lines, int32_t comments);
 
 #endif
